Active song is now in player only not in queue, other stuff
Daphne:
* Catch Application.shutdown signal and call quit() to cleanly shutdown
* Use player "song" property to update coverDisplay

HistoryColumnView.pop: Properly handle empty item list

PlayQueue:
* Move some logic to SongColumnView
* Rename _songColumnView to songColumnView and make public
* Remove songCount and currentSong properties (handled by SongColumnView and Player)
* Separate database update logic from onShuffleButtonClicked() to reindexQueueTable()
* Remove start() and stop() methods (Player handles that now)
* Add pop() and push() methods for popping a song off the top of the queue or pushing one back on
* insert() now handles IDs properly when inserting items
* Remove next(), prev(), and remove() methods since Player and SongColumnView handles this now

Player:
* Current song is now in player only (not in queue as well)
* Uses playQueue.songColumnView to watch queue song count
* canGoPrevious is now set properly
* play() now takes an optional song argument
* Update Player and PlayQueue logic
* stop() method now takes a PushAction enum to determine which list the song is pushed on to (PlayQueue, History, or None)

SongColumnView: Add songCount property for watching for song count changes

+++ source/history_column_view.d
@@ -123,12 +123,14 @@ class HistoryColumnView : SongColumnView
   alias pop = SongColumnView.pop;
 
   /**
-   * Pop the current song off of the history. Removes it from the history and returns it.
+   * Pop the song off of the end of the history. Removes it from the history and returns it.
    * Returns: Song popped off of the end of the history or null if none
    */
   LibrarySong pop()
   {
     auto item = cast(HistoryColumnViewItem)super.pop;
+    if (!item)
+      return null;
 
     try
       _daphne.playQueue.dbConn.executeSql("DELETE FROM history WHERE id=" ~ item.id.to!string);
